Split Slate editor into ScriptureTextPanelSlate, fixed double scroll bars, fixed error changing chapter on selection
