+++ react-electron-poc/src/renderer/components/panels/Panels.ts
@@ -5,6 +5,7 @@ import { TextPanel } from './TextPanels/TextPanel';
 import { HtmlTextPanel } from './TextPanels/HtmlTextPanel';
 import { EditableHtmlTextPanel } from './TextPanels/EditableHtmlTextPanel';
 import { ScriptureTextPanel } from './TextPanels/ScriptureTextPanel';
+import { ScriptureTextPanelSlate } from './TextPanels/ScriptureTextPanelSlate';
 
 // TODO: Consider doing something a bit different with this https://stackoverflow.com/questions/29722270/is-it-possible-to-import-modules-from-all-files-in-a-directory-using-a-wildcard
 /** All available panels for use in dockviews */
@@ -14,6 +15,7 @@ export const Panels = {
     HtmlTextPanel,
     EditableHtmlTextPanel,
     ScriptureTextPanel,
+    ScriptureTextPanelSlate,
 };
 export default Panels;
